Base comp ser (#1089)
* Added serializable_base as the base serializable class.  Serializable and BaseComponent now inherit from serializable_base.  This allows different serialize_impl<> templates for serializable and BaseComponent, giving more control over Component serialization.

* Fixed bug for --run-mode=init. Fixes #1087.

* Made serialization tests run with --run-mode=init since they test everything during the construct phase.  This will test that --run-mode=init operates correctly.

* Added [[maybe_unused]] to __Eli_Derived_Level at suggestion of @ldalessa
