+++ tests/refFiles/test_Serialization_default.out
@@ -1,2 +1,2 @@
 WARNING: Building component "Component0" with no links assigned.
-Simulation is complete, simulated time: 1 us
+Simulation is complete, simulated time: 0 s
