+++ tests/refFiles/test_Serialization_handler.out
@@ -63,4 +63,4 @@ metadata value: 8.9
 internal value: 73
 parameter value: 17
 metadata value: 8.9
-Simulation is complete, simulated time: 1 us
+Simulation is complete, simulated time: 0 s
